LRender/requirements.txt
numpy~=1.26.1
Pillow~=10.1.0
Cython~=3.0.4
PyQt5~=5.15.9